refactor(cli): remove old cli components
