+++ .ci/xdp/integration-test.sh
@@ -34,7 +34,7 @@ spec:
     ports:
     - containerPort: 7777
       hostPort: 7777
-    args: ["proxy", "--to", "${server_ip}:8078", "--service.udp.xdp"]
+    args: ["--service.udp", "--provider.static.endpoints=${server_ip}:8078", "--service.udp.xdp"]
     securityContext:
       capabilities:
         add:
