+++ Cargo.toml
@@ -252,7 +252,6 @@ parking_lot = "0.12.3"
 schemars = { version = "0.8.21", features = ["bytes", "url"] }
 url = { version = "2.5.3", features = ["serde"] }
 xdp = "0.7.0"
-#xdp = { path = "../xdp" }
 
 [workspace.lints.clippy]
 undocumented_unsafe_blocks = "deny"
