+++ benches/shared.rs
@@ -324,27 +324,31 @@ impl QuilkinLoop {
 
         let thread = spawn("quilkin", move || {
             let runtime = tokio::runtime::Runtime::new().unwrap();
-            let config = Arc::new(quilkin::Config::default_non_agent());
+            let config = Arc::new(quilkin::Config::default());
             config.dyn_cfg.clusters().unwrap().modify(|clusters| {
                 clusters.insert_default(
                     [quilkin::net::endpoint::Endpoint::new(endpoint.into())].into(),
                 );
             });
 
-            let proxy = quilkin::cli::Proxy {
-                port,
-                qcmp_port: runtime
-                    .block_on(quilkin::test::available_addr(
-                        quilkin::test::AddressType::Random,
-                    ))
-                    .port(),
-                ..<_>::default()
-            };
+            let proxy = quilkin::cli::Service::builder()
+                .udp()
+                .udp_port(port)
+                .qcmp()
+                .qcmp_port(
+                    runtime
+                        .block_on(quilkin::test::available_addr(
+                            quilkin::test::AddressType::Random,
+                        ))
+                        .port(),
+                );
 
             runtime.block_on(async move {
                 proxy
-                    .run(config, Default::default(), None, shutdown_rx)
+                    .spawn_services(&config, &shutdown_rx)
+                    .unwrap()
                     .await
+                    .unwrap()
                     .unwrap();
             });
         });
