+++ crates/agones/src/pod.rs
@@ -31,7 +31,13 @@ mod tests {
         let client = Client::new().await;
 
         let pods: Api<Pod> = client.namespaced_api();
-        let cmds = ["proxy", "--to", "127.0.0.1:0"].map(String::from).to_vec();
+        let cmds = [
+            "--service.udp",
+            "--provider.static.endpoints",
+            "127.0.0.1:9999",
+        ]
+        .map(String::from)
+        .to_vec();
         let pod = Pod {
             metadata: ObjectMeta {
                 generate_name: Some("quilkin-".into()),
