+++ crates/test/Cargo.toml
@@ -37,4 +37,6 @@ tempfile.workspace = true
 tokio = { workspace = true, features = ["macros"] }
 tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
+
+[target.'cfg(target_os = "linux")'.dependencies]
 xdp = { workspace = true, features = ["__debug"] }
