+++ docs/preprocessor.sh
@@ -20,12 +20,6 @@ cargo run -q --manifest-path ../Cargo.toml -- -q generate-config-schema -o ../ta
 
 # Output all the command line help
 cargo run -q --manifest-path ../Cargo.toml &> ../target/quilkin.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- proxy --help &> ../target/quilkin.proxy.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- manage --help &> ../target/quilkin.manage.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- manage agones --help &> ../target/quilkin.manage.agones.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- manage file --help &> ../target/quilkin.manage.file.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- relay --help &> ../target/quilkin.relay.commands || true
-cargo run -q --manifest-path ../Cargo.toml -- agent --help &> ../target/quilkin.agent.commands || true
 
 # Credit: https://github.com/rust-lang/mdBook/issues/1462#issuecomment-778650045
 jq -M -c .[1] <&0
