--- examples/agones-xonotic-sidecar/sidecar-compress.yaml
@@ -1,77 +1,0 @@
-#
-# Copyright 2021 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: quilkin-config-compress
-data:
-  quilkin.yaml: |
-    version: v1alpha1
-    filters:
-      - name: quilkin.filters.compress.v1alpha1.Compress
-        config:
-          on_read: DECOMPRESS
-          on_write: COMPRESS
-          mode: SNAPPY
-    clusters:
-      - endpoints:
-        - address: 127.0.0.1:26000
----
-apiVersion: "agones.dev/v1"
-kind: Fleet
-metadata:
-  name: xonotic-sidecar-compress
-spec:
-  replicas: 2
-  template:
-    spec:
-      container: xonotic
-      ports:
-        - name: default
-          containerPort: 26001
-          container: quilkin
-      health:
-        initialDelaySeconds: 30
-        periodSeconds: 60
-      template:
-        metadata:
-          annotations:
-            prometheus.io/scrape: "true"
-            prometheus.io/path: /metrics
-            prometheus.io/port: "8000"
-        spec:
-          containers:
-            - name: xonotic
-              image: us-docker.pkg.dev/agones-images/examples/xonotic-example:1.2
-            - name: quilkin
-              image: us-docker.pkg.dev/quilkin/release/quilkin:0.9.0
-              args:
-                - proxy
-                - --port=26001
-              volumeMounts:
-                - name: config
-                  mountPath: "/etc/quilkin"
-              livenessProbe:
-                httpGet:
-                  path: /live
-                  port: 8000
-                initialDelaySeconds: 3
-                periodSeconds: 2
-          volumes:
-            - name: config
-              configMap:
-                name: quilkin-config-compress
