+++ examples/agones-xonotic-sidecar/sidecar.yaml
@@ -42,7 +42,10 @@ spec:
               image: us-docker.pkg.dev/agones-images/examples/xonotic-example:1.2
             - name: quilkin
               image: us-docker.pkg.dev/quilkin/release/quilkin:0.9.0
-              args: ["proxy", "--port", "26001", "--to", "127.0.0.1:26000"]
+              args:
+                - --service.udp
+                - --service.udp.port=26001
+                - --provider.static.endpoints=127.0.0.1:26000
               livenessProbe:
                 httpGet:
                   path: /live
