+++ examples/agones-xonotic-xds/proxy-pool.yaml
@@ -43,7 +43,9 @@ spec:
       containers:
         - name: quilkin
           image: us-docker.pkg.dev/quilkin/release/quilkin:0.9.0
-          args: ["proxy", "--management-server", "http://quilkin-manage-agones:80"]
+          args:
+            - --service.udp
+            - --provider.xds.endpoints=http://quilkin-manage-agones
           env:
             - name: RUST_LOG
               value: info #,quilkin=trace
