+++ src/components.rs
@@ -15,10 +15,7 @@
  */
 
 pub mod admin;
-pub mod agent;
-pub mod manage;
 pub mod proxy;
-pub mod relay;
 
 /// Args common across all components
 pub struct RunArgs<T> {
