+++ src/components/proxy/sessions.rs
@@ -578,7 +578,7 @@ mod tests {
         let (pending_sends, _srecv) = crate::net::queue(1).unwrap();
         (
             SessionPool::new(
-                Arc::new(Config::default_agent()),
+                Arc::new(Config::default()),
                 vec![pending_sends.clone()],
                 Arc::new(BufferPool::default()),
             ),
