+++ src/config.rs
@@ -46,7 +46,6 @@ pub use self::{
 mod config_type;
 mod error;
 pub mod providers;
-pub mod providersv2;
 mod serialization;
 mod slot;
 pub mod watch;
@@ -906,10 +905,18 @@ impl Config {
         crate::metrics::apply_clusters(clusters);
     }
 
-    pub fn default_agent() -> Self {
+    #[inline]
+    pub fn id(&self) -> String {
+        String::clone(&self.dyn_cfg.id.load())
+    }
+}
+
+impl Default for Config {
+    fn default() -> Self {
         let mut typemap = default_typemap();
         insert_default::<FilterChain>(&mut typemap);
         insert_default::<ClusterMap>(&mut typemap);
+        insert_default::<DatacenterMap>(&mut typemap);
         insert_default::<Agent>(&mut typemap);
 
         Self {
@@ -919,26 +926,6 @@ impl Config {
                 typemap,
             },
         }
-    }
-
-    pub fn default_non_agent() -> Self {
-        let mut typemap = default_typemap();
-        insert_default::<FilterChain>(&mut typemap);
-        insert_default::<ClusterMap>(&mut typemap);
-        insert_default::<DatacenterMap>(&mut typemap);
-
-        Self {
-            dyn_cfg: DynamicConfig {
-                id: default_id(),
-                version: Version::default(),
-                typemap,
-            },
-        }
-    }
-
-    #[inline]
-    pub fn id(&self) -> String {
-        String::clone(&self.dyn_cfg.id.load())
     }
 }
 
@@ -1325,4 +1312,19 @@ pub enum AddrKind {
     Ipv4,
     Ipv6,
     Any,
+}
+
+impl clap::ValueEnum for crate::config::AddrKind {
+    fn value_variants<'a>() -> &'a [Self] {
+        &[Self::Ipv4, Self::Ipv6, Self::Any]
+    }
+
+    fn to_possible_value(&self) -> Option<clap::builder::PossibleValue> {
+        use clap::builder::PossibleValue as pv;
+        Some(match self {
+            Self::Ipv4 => pv::new("v4"),
+            Self::Ipv6 => pv::new("v6"),
+            Self::Any => pv::new("any"),
+        })
+    }
 }
